Update Delete, delete Alert

+++ src/components/Element.js
@@ -1,5 +1,6 @@
 import React from 'react'
 import { Link } from 'react-router-dom'
+import Swal from 'sweetalert2'
 
 // Redux
 import { useDispatch } from 'react-redux'
@@ -12,8 +13,21 @@ const Element = ({element}) =>{
     const buttonDelete = id => {
         //answer user
 
-        //go to Action
-        dispatch( deleteElementAction(id) )
+        Swal.fire({
+            title: 'Are you sure?',
+            text: "You won't be able to revert this!",
+            icon: 'warning',
+            showCancelButton: true,
+            confirmButtonColor: '#3085d6',
+            cancelButtonColor: '#d33',
+            confirmButtonText: 'Yes, delete it!'
+          }).then((result) => {
+            if (result.value) {
+                //go to Action
+                dispatch( deleteElementAction(id) )
+            }
+          })
+
     }
 
     return(
